requirements.txt
docopt
flask
flask-talisman
requests
pytz
gunicorn
BeautifulSoup4
flask_sqlalchemy
Flask-Script
